Extract engine add-audio operation
add_audio has distinct media attachment and optional mixing behavior that should not live in the engine facade. This moves it into engine_audio_ops.py while re-exporting add_audio from mcp_video.engine for compatibility.

Constraint: Preserve audio replacement, mixing, delay, fade, probing, and EditResult behavior.

Rejected: Move extract_audio and normalize_audio in the same PR | those are separate operation families with different contracts.

Confidence: high

Scope-risk: narrow

Directive: Keep engine audio operations split by behavior rather than creating one oversized audio module.

Tested: ruff check --fix mcp_video/engine.py mcp_video/engine_audio_ops.py

Tested: /opt/homebrew/bin/python3 add_audio re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py::TestAddAudio tests/test_server.py::TestVideoAddAudioTool tests/test_e2e.py::TestTikTokWorkflow::test_tiktok_workflow -q --tb=short

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine.py tests/test_e2e.py tests/test_server.py -q --tb=short
